# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/v1/caracteristicasunidadconstruccion/convencional": {
            "get": {
                "tags": [
                    "Construccion"
                ],
                "summary": "Obtener información de contrucciones convencionales",
                "operationId": "f3eb283d5cf6d0a97c320311381e3878",
                "responses": {
                    "200": {
                        "description": "Información de la construccion convencional",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la respuesta",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje de la respuesta",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Datos de la construccion",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ConvencionalCollection"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la respuesta",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Construccion"
                ],
                "summary": "Almacenar construcciones convencionales",
                "operationId": "a002d217a44f94f2759b163b24dca145",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidarConstruccionesFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contruccion creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Estado de la peticion",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Estado de la peticion",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/caracteristicasunidadconstruccion/no-convencional": {
            "get": {
                "tags": [
                    "Construccion"
                ],
                "summary": "Obtener información de contrucciones no convencionales",
                "operationId": "752eccd8eac531d2244ff623839a1858",
                "responses": {
                    "200": {
                        "description": "Información de la construccion no convencional",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la respuesta",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje de la respuesta",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Datos de la construccion",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NoConvencionalCollection"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la respuesta",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Construccion"
                ],
                "summary": "Almacenar construcciones no convencionales",
                "operationId": "f6015cd6b1ea26f523a427e1be9723c9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidarConstruccionesNoConvencionalesFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contruccion creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Estado de la peticion",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Estado de la peticion",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/predio/{id}": {
            "get": {
                "tags": [
                    "Predio"
                ],
                "summary": "Obtener información de un predio por ID",
                "operationId": "ee083f75f1bce643aa73942d21b55b76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del predio a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del predio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la respuesta",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje de la respuesta",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Datos del predio",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PredioResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la respuesta",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidarConstruccionesFormRequest": {
                "properties": {
                    "construcciones": {
                        "description": "Lista de construcciones",
                        "type": "array",
                        "items": {
                            "properties": {
                                "caracteristicasunidadconstruccion": {
                                    "required": [
                                        "identificador",
                                        "tipo_construccion",
                                        "tipo_dominio",
                                        "tipo_unidad_construccion",
                                        "tipo_planta",
                                        "total_plantas",
                                        "total_habitaciones",
                                        "total_banios",
                                        "total_locales",
                                        "anio_construccion",
                                        "uso",
                                        "avaluo_unidad_construccion",
                                        "area_construida",
                                        "area_privada_construida",
                                        "fin_vida_util_version",
                                        "espacio_de_nombres",
                                        "local_id",
                                        "observaciones",
                                        "calificacionconvencional"
                                    ],
                                    "properties": {
                                        "identificador": {
                                            "description": "Identificador de la construcción",
                                            "type": "string",
                                            "maxLength": 20
                                        },
                                        "tipo_construccion": {
                                            "description": "ID del tipo de construcción",
                                            "type": "integer"
                                        },
                                        "tipo_dominio": {
                                            "description": "ID del tipo de dominio",
                                            "type": "integer"
                                        },
                                        "tipo_unidad_construccion": {
                                            "description": "ID del tipo de unidad de construcción",
                                            "type": "integer"
                                        },
                                        "tipo_planta": {
                                            "description": "ID del tipo de planta",
                                            "type": "integer"
                                        },
                                        "total_plantas": {
                                            "description": "Total de plantas",
                                            "type": "integer"
                                        },
                                        "total_habitaciones": {
                                            "description": "Total de habitaciones",
                                            "type": "integer"
                                        },
                                        "total_banios": {
                                            "description": "Total de baños",
                                            "type": "integer"
                                        },
                                        "total_locales": {
                                            "description": "Total de locales",
                                            "type": "integer"
                                        },
                                        "anio_construccion": {
                                            "description": "Año de construcción",
                                            "type": "integer"
                                        },
                                        "uso": {
                                            "description": "ID del tipo de uso de construcción",
                                            "type": "integer"
                                        },
                                        "avaluo_unidad_construccion": {
                                            "description": "Avalúo de la unidad de construcción",
                                            "type": "string"
                                        },
                                        "area_construida": {
                                            "description": "Área construida",
                                            "type": "number"
                                        },
                                        "area_privada_construida": {
                                            "description": "Área privada construida",
                                            "type": "number"
                                        },
                                        "fin_vida_util_version": {
                                            "description": "Fin de vida útil (versión)",
                                            "type": "string"
                                        },
                                        "espacio_de_nombres": {
                                            "description": "Espacio de nombres",
                                            "type": "string",
                                            "maxLength": 255
                                        },
                                        "local_id": {
                                            "description": "ID del local",
                                            "type": "string",
                                            "maxLength": 255
                                        },
                                        "observaciones": {
                                            "description": "Observaciones",
                                            "type": "string"
                                        },
                                        "calificacionconvencional": {
                                            "description": "Calificación convencional",
                                            "type": "array",
                                            "items": {
                                                "required": [
                                                    "tipo_calificar",
                                                    "total_calificacion",
                                                    "grupocalificacion"
                                                ],
                                                "properties": {
                                                    "tipo_calificar": {
                                                        "description": "ID del tipo de calificación",
                                                        "type": "integer"
                                                    },
                                                    "total_calificacion": {
                                                        "description": "Total de calificación",
                                                        "type": "integer"
                                                    },
                                                    "grupocalificacion": {
                                                        "description": "Grupo de calificación",
                                                        "type": "array",
                                                        "items": {
                                                            "required": [
                                                                "clase_calificacion",
                                                                "conservacion",
                                                                "subtotal",
                                                                "objetoconstruccion"
                                                            ],
                                                            "properties": {
                                                                "clase_calificacion": {
                                                                    "description": "ID de la clase de calificación",
                                                                    "type": "integer"
                                                                },
                                                                "conservacion": {
                                                                    "description": "ID del estado de conservación",
                                                                    "type": "integer"
                                                                },
                                                                "subtotal": {
                                                                    "description": "Subtotal de calificación",
                                                                    "type": "integer"
                                                                },
                                                                "objetoconstruccion": {
                                                                    "description": "Objeto de construcción",
                                                                    "type": "array",
                                                                    "items": {
                                                                        "required": [
                                                                            "tipo_objeto_construccion",
                                                                            "puntos"
                                                                        ],
                                                                        "properties": {
                                                                            "tipo_objeto_construccion": {
                                                                                "description": "ID del tipo de objeto de construcción",
                                                                                "type": "integer"
                                                                            },
                                                                            "puntos": {
                                                                                "description": "Puntos de objeto de construcción",
                                                                                "type": "integer"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}